Split, match and cleanup stuff
