+++ config/sly1.yaml
@@ -139,14 +139,14 @@ segments:
     - [0x60678, asm, P2/freeze]
     - [0x60f88, asm, P2/frzg]
     - [0x61070, c, P2/game]
-    - [0x61e80, asm, P2/geom]
-    #- [0x, asm, P2/gifs]
-    #- [0x, asm, P2/glbs]
-    #- [0x, asm, P2/glob]
-    #- [0x, asm, P2/gomer]
+    - [0x61e80, c, P2/geom]
+    - [0x621b8, c, P2/gifs]
+    - [0x629c0, c, P2/glbs]
+    - [0x63810, c, P2/glob]
+    - [0x68d68, c, P2/gomer]
     - [0x69130, c, P2/gs]
-    - [0x6ac58, asm, P2/hide]
-    #- [0x, asm, P2/hnd]
+    - [0x6ac58, c, P2/hide]
+    - [0x6ba18, asm, P2/hnd] # Might be something else
     #- [0x, asm, P2/ik]
     #- [0x, asm, P2/jack]
     #- [0x, asm, P2/jlo]
